requirements.txt
# Core dependencies for next-location prediction research project
# Python 3.9+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Experiment Tracking & Visualization
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress & Logging
tqdm>=4.65.0
colorlog>=6.7.0
pyyaml>=6.0

# Utilities
scikit-learn>=1.2.0
h5py>=3.8.0

# Code Quality (optional, for development)
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Reproducibility
wandb>=0.15.0  # Optional: for cloud experiment tracking
